Refactor account
