+++ assets/src/components/account/groups/queries.ts
@@ -1,0 +1,96 @@
+import { gql } from 'apollo-boost'
+
+import { PageInfo } from '../../graphql/base'
+import { GroupFragment, GroupMemberFragment } from '../../graphql/users'
+
+export const GROUPS_Q = gql`
+  query Groups($q: String, $cursor: String) {
+    groups(q: $q, first: 20, after: $cursor) {
+      pageInfo {
+        ...PageInfo
+      }
+      edges {
+        node {
+          ...GroupFragment
+        }
+      }
+    }
+  }
+  ${PageInfo}
+  ${GroupFragment}
+`
+
+export const SEARCH_GROUPS = gql`
+  query SearchGroups($q: String, $cursor: String) {
+    groups(q: $q, after: $cursor, first: 5) {
+      pageInfo {
+        ...PageInfo
+      }
+      edges {
+        node {
+          ...GroupFragment
+        }
+      }
+    }
+  }
+  ${PageInfo}
+  ${GroupFragment}
+`
+
+export const GROUP_MEMBERS = gql`
+  query GroupMembers($cursor: String, $id: ID!) {
+    groupMembers(groupId: $id, after: $cursor, first: 20) {
+      pageInfo { ...PageInfo }
+      edges {
+        node { ...GroupMemberFragment }
+      }
+    }
+  }
+  ${PageInfo}
+  ${GroupMemberFragment}
+`
+
+export const CREATE_GROUP_MEMBERS = gql`
+  mutation CreateGroupMember($groupId: ID!, $userId: ID!) {
+    createGroupMember(groupId: $groupId, userId: $userId) {
+      ...GroupMemberFragment
+    }
+  }
+  ${GroupMemberFragment}
+`
+
+export const DELETE_GROUP_MEMBER = gql`
+  mutation DeleteMember($groupId: ID!, $userId: ID!) {
+    deleteGroupMember(groupId: $groupId, userId: $userId) {
+      ...GroupMemberFragment
+    }
+  }
+  ${GroupMemberFragment}
+`
+
+export const CREATE_GROUP = gql`
+  mutation CreateGroup($attributes: GroupAttributes!) {
+    createGroup(attributes: $attributes) {
+      ...GroupFragment
+    }
+  }
+  ${GroupFragment}
+`
+
+export const UPDATE_GROUP = gql`
+  mutation UpdateGroup($id: ID!, $attributes: GroupAttributes!) {
+    updateGroup(groupId: $id, attributes: $attributes) {
+      ...GroupFragment
+    }
+  }
+  ${GroupFragment}
+`
+
+export const DELETE_GROUP = gql`
+  mutation DeleteGroup($id: ID!) {
+    deleteGroup(groupId: $id) {
+      ...GroupFragment
+    }
+  }
+  ${GroupFragment}
+`
