+++ assets/src/components/account/roles/RoleEdit.tsx
@@ -4,10 +4,10 @@ import { useMemo, useState } from 'react'
 import uniqWith from 'lodash/uniqWith'
 import isEqual from 'lodash/isEqual'
 
-import { UPDATE_ROLE } from '../queries'
-
-import { Actions } from '../Actions'
+import { Actions } from '../../utils/Actions'
 import { sanitize } from '../utils'
+
+import { UPDATE_ROLE } from './queries'
 
 import RoleForm from './RoleForm'
