--- assets/src/components/account/types.ts
@@ -1,6 +1,0 @@
-export const PermissionTypes = {
-  READ: 'can view components',
-  CONFIGURE: 'can edit helm/terraform configuration',
-  DEPLOY: 'can create/approve deployments',
-  OPERATE: 'can delete pods, export logs, and other operational tasks',
-}
